Inserindo teste de integração

+++ backend/src/usecases/delete-currency.usecase.ts
@@ -1,8 +1,14 @@
 import { ICurrencyRepository } from '../domain/repository/ICurrencyRepository';
+import { BadRequestException } from '@nestjs/common';
 
 export class DeleteCurrencyUseCase {
   constructor(private readonly currencyRepository: ICurrencyRepository) {}
-  exec(alias: string) {
+  async exec(alias: string) {
+    const getCurrency = await this.currencyRepository.findByAlias(alias);
+
+    if (!getCurrency) {
+      throw new BadRequestException('Currency not found');
+    }
     return this.currencyRepository.delete(alias);
   }
 }
